Use CRUD functions in AccountType
